Gracefully handle sigINT and sigTERM
Stop and remove all started containers upon receiving termination
request.

Because termination handler runs in another thread (as decribed in
http://hackage.haskell.org/package/unix-2.7.2.2/docs/System-Posix-Signals.html#v:installHandler)
and needs a bit of time to stop all containers, we need a way to tell
main thread to wait until cleanup is finished. Otherwise, main thread
will exit too early, and termination handler thread will be killed, not
able to finish cleanup. Communication with main thread is done using
`MVar` as a flag, because `takeMVar` will wait until `MVar` is filled
with value. For better clarity this `MVar` is aliased as `ExitFlag`.

Closing websocket connection poses another challenge, because
termination handler needs access to connection to be able to close it,
so connection now lives in global `State`.

`connectToMessenger` behavior is altered in a way that it recognzes that
connection is being closed, and doesn't attempt to reconnect.

Error handling in the `app` function is removed, because anyway `app`
runs inside `connectToMessenger`, so it can catch errors, and it turned
out that `ConnectionClosed` exceptions have to be handled there to stop
reconnection attempts.