Report when some long-running operations are done
This is needed by integration tests, which look at the output produced
by `hex-exe` and wait until requested operation is complete.

Also because of the tests we need to set stdout buffering to
`LineBuffering`, because otherwise integration tests aren't able to read
output in timely fashion (default setting is `BlockBuffering`, which
means that until block of some size is not filled completely,
integration test will not receive *anything* that `hex-exe` outputs).