Configure messenger host via Hexfile
Hex could be run in two environments: inside Docker container (in
production) and on the host system directly (usually in development, but
also on CI). When run inside Docker container, it is provided with
hostname resolution by Docker, and host `messenger` would resolve to
correct Arnaux container, but there is no such hostname resolution when
Hex is running on the host system.

To get around this issue I've added a new key to messenger definition
which allows us to specify websocket host to which Hex should connect.