Track running services
This module is intended as replacement for `State`, with certain design
principle in mind: to make invalid states irrepresentable.

This is achieved by these strategies:
- Using `Set` internally to store list of services. Set is hashed by
container id from Docker, because this value uniquely identifies running
service.
- Storing all information about service in one piece, instead of
separate lists of "request queue" and "container ids". This makes it
impossible for service to be in both lists at the same time.