Reword API for RunningServices
I feel that it's better to operate in terms of "started/running"
services, instead of "request/fulfill", because it's actually not a
"request" to run a service, but rather storing information about already
started Docker container. And we aren't "fulfilling" request, we simply
storing information that we got confirmation (checkin message) from
started service, which means it's operating correctly.