Validations for HexFile
It was possible to parse and create invalid `HexFile` - for example,
messenger definition could be missing, or definitions for services
listed in `init-sequence`. These errors were handled later at runtime.

This change pushes validation of `HexFile` to parsing stage, removing
the need to worry about missing services at runtime.

This is achieved mainly by altering `HexFile` data structure in a way
that makes it impossible to construct invalid `HexFile`, and by
providing smart parsers that use information from previous stages of
parsing to either parse successfully or fail with error.